pyproject.toml
[project]
name = "houseagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "openai>=2.3.0",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-ulid>=3.1.0",
    "structlog>=24.4.0",
]

[tool.setuptools]
packages = ["houseagent"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
